backend/requirements-simple.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
deepface==0.0.79
opencv-python-headless==4.8.1.78
pillow==10.1.0
requests==2.31.0
